requirements.txt
asgiref==3.9.2
Django==5.2.7
dnspython==2.8.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
redis==6.4.0
sqlparse==0.5.3
